+++ changelog.md
@@ -1,1 +1,3 @@
-* Agregado la posibilidad de leer el contenido de portapapeles incluido portapapeles de más de 1024 caracteres.
+* Agregadas  estadísticas del texto seleccionado
+* Agregada la posibilidad de copiar partes de un texto seleccionado.
+* Añadida compatibilidad con NVDA 2023.1